Refactor IbisDocApp code into the respective objects
