+++ core/src/main/java/nl/nn/adapterframework/doc/objects/AClass.java
@@ -17,8 +17,13 @@ package nl.nn.adapterframework.doc.objects;
 
 import nl.nn.adapterframework.doc.IbisDoc;
 import nl.nn.adapterframework.doc.IbisDocRef;
+import org.springframework.core.annotation.AnnotationUtils;
 
+import java.lang.reflect.Method;
 import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.Map;
+import java.util.TreeSet;
 
 /**
  * Represents the Class object for the IbisDoc application
@@ -32,13 +37,13 @@ public class AClass {
     private String javadocLink;
     private ArrayList<String> superClasses;
     private ArrayList<AMethod> methods;
+    private String referredClassName = "";
 
-    public AClass(String name, String packageName, String javadocLink) {
+
+    public AClass(String name) {
         this.name = name;
-        this.packageName = packageName;
-        this.javadocLink = javadocLink;
         this.superClasses = new ArrayList<>();
-        this.methods = new ArrayList<AMethod>();
+        this.methods = new ArrayList<>();
     }
     /**
      * Get the superclasses of a certain class.
@@ -47,7 +52,7 @@ public class AClass {
      * @return An ArrayList containing all the superclasses with priority given to
      *         them
      */
-    public void addSuperClasses(String referredClassName) {
+    public void setSuperclasses(String referredClassName) {
         if (!referredClassName.isEmpty()) {
             superClasses.add(referredClassName);
         }
@@ -67,6 +72,67 @@ public class AClass {
         }
     }
 
+    /**
+     * Add the methods to the class object.
+     *
+     * @param beanProperties - The properties of a class
+     * @param newClass       - The class object we have to add the methods to
+     * @return the AClass with the added AMethods
+     */
+    public void setMethods(Map<String, Method> beanProperties, AClass newClass) {
+        Iterator<String> iterator = new TreeSet<>(beanProperties.keySet()).iterator();
+        setReferredClassName("");
+        while (iterator.hasNext()) {
+
+            // Get the method
+            String property = iterator.next();
+            Method method = beanProperties.get(property);
+            AMethod aMethod = new AMethod(property);
+
+            // Get the IbisDocRef
+            IbisDocRef reference = AnnotationUtils.findAnnotation(method, IbisDocRef.class);
+
+            // Get the IbisDoc values from the annotations above the method
+            IbisDoc ibisDoc = AnnotationUtils.findAnnotation(method, IbisDoc.class);
+
+            // Check for whether the method (attribute) is deprecated
+            Deprecated deprecated = AnnotationUtils.findAnnotation(method, Deprecated.class);
+            boolean isDeprecated = deprecated != null;
+
+            String order = "";
+            String originalClassName = "";
+
+            // If there is an IbisDocRef for the method, get the IbisDoc of the referred method
+            if (reference != null) {
+                order = reference.value()[0];
+                ibisDoc = aMethod.getIbisDocRef(reference.value()[1], method);
+                setReferredClassName(aMethod.getReferredClassName());
+                originalClassName = aMethod.getReferredClassName();
+            }
+
+            // If there is an IbisDoc for the method, add the method and it's IbisDoc values to the class object
+            if (ibisDoc != null) {
+                String[] ibisdocValues = ibisDoc.value();
+                String[] values = aMethod.getValues(ibisdocValues);
+
+                // This is done for @IbisDoc use instead of @IbisDocRef
+                if (order.isEmpty()) order = values[2];
+
+                // This is done for @IbisDoc use instead of @IbisDocRef
+                if (originalClassName.isEmpty()) originalClassName = method.getDeclaringClass().getSimpleName();
+
+                aMethod.setOriginalClassName(originalClassName);
+                aMethod.setDescription(values[0]);
+                aMethod.setDefaultValue(values[1]);
+                aMethod.setOrder(Integer.parseInt(order));
+                aMethod.setDeprecated(isDeprecated);
+                aMethod.setReferredClassName(getReferredClassName());
+
+                newClass.addMethod(aMethod);
+            }
+        }
+        newClass.setSuperclasses(getReferredClassName());
+    }
 
     public void addMethod(AMethod method) {
         methods.add(method);
@@ -76,19 +142,39 @@ public class AClass {
         return name;
     }
 
+    public void setName(String name) {
+        this.name = name;
+    }
+
     public String getPackageName() {
         return packageName;
     }
-    
+
+    public void setPackageName(String packageName) {
+        this.packageName = packageName;
+    }
+
     public String getJavadocLink() {
-    	return javadocLink;
+        return javadocLink;
+    }
+
+    public void setJavadocLink(String javadocLink) {
+        this.javadocLink = javadocLink;
+    }
+
+    public ArrayList<String> getSuperClasses() {
+        return superClasses;
     }
 
     public ArrayList<AMethod> getMethods() {
         return methods;
     }
-    
-    public ArrayList<String> getSuperClasses() {
-    	return superClasses;
+
+    public String getReferredClassName() {
+        return this.referredClassName;
+    }
+
+    public void setReferredClassName(String referredClassName) {
+        this.referredClassName = referredClassName;
     }
 }
