+++ core/src/main/java/nl/nn/adapterframework/doc/objects/AFolder.java
@@ -15,7 +15,12 @@
 */
 package nl.nn.adapterframework.doc.objects;
 
+import nl.nn.adapterframework.doc.IbisDocPipe;
+
+import java.lang.reflect.Method;
 import java.util.ArrayList;
+import java.util.Map;
+import java.util.TreeSet;
 
 /**
  * Represents the Folder object for the IbisDoc application
@@ -29,9 +34,32 @@ public class AFolder {
 
     public AFolder(String name) {
         this.name = name;
-        this.classes = new ArrayList<AClass>();
+        this.classes = new ArrayList<>();
     }
 
+    /**
+     * Add classes to the folder object.
+     *
+     * @param groups - Contains all information
+     * @param folder - The folder object we have to add the classes to
+     */
+    public void setClasses(Map<String, TreeSet<IbisBean>> groups, AFolder folder) {
+        for (IbisBean ibisBean : groups.get(folder.getName())) {
+            Map<String, Method> beanProperties = IbisDocPipe.getBeanProperties(ibisBean.getClazz());
+            if (!beanProperties.isEmpty()) {
+                AClass aClass = new AClass(ibisBean.getName());
+
+                // Get the javadoc link for the class
+                String javadocLink = ibisBean.getClazz().getName().replaceAll("\\.", "/");
+
+                // Create a new class and add the methods (attributes) to it, then add it to the folder object
+                aClass.setJavadocLink(javadocLink);
+                aClass.setPackageName(ibisBean.getClazz().getName());
+                aClass.setMethods(beanProperties, aClass);
+                folder.addClass(aClass);
+            }
+        }
+    }
     public void addClass(AClass clazz) {
         this.classes.add(clazz);
     }
